+++ backend/src/main/java/backend/mulkkam/device/controller/DeviceController.java
@@ -1,0 +1,24 @@
+package backend.mulkkam.device.controller;
+
+import backend.mulkkam.device.dto.CreateDeviceRequest;
+import backend.mulkkam.device.service.DeviceService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RequestMapping("/devices")
+@RestController
+public class DeviceController {
+
+    private final DeviceService deviceService;
+
+    @PostMapping()
+    public ResponseEntity<Void> create(@RequestBody CreateDeviceRequest createDeviceRequest) {
+        deviceService.create(createDeviceRequest, 1L);
+        return ResponseEntity.ok().build();
+    }
+}
